Fix crash at startup due to kGCMChannelStatus not being registered
This boolean preference has now been removed from the codebase, causing
Brave to crash on startup since the SetDefaultPrefValue() call to set
its default value will now be operating on an unregistered preference.

If we want to keep the migration code working, we need to register the
preference ourselves now, as well as to make sure it gets cleared once
the migration step has been completed.

commit 7f84e6bdce3395eef254aec617bc76dff41b3a8f
Author: Tibor Goldschwendt <tiborg@chromium.org>
Date:   Fri Feb 26 22:21:35 2021 +0000

    Cleanup prefs deprecated in 1/2020 and 2/2020